[Fix] 비용 계산 로직 서비스로 이동 및 멤버 삭제 버그 수정

+++ src/main/java/com/llm_ops/demo/gateway/dto/GatewayChatUsage.java
@@ -1,5 +1,7 @@
 package com.llm_ops.demo.gateway.dto;
 
+import java.math.BigDecimal;
+
 /**
  * 게이트웨이 채팅 API 응답에 포함될 리소스 사용량 정보를 위한 DTO 레코드입니다.
  * '예산 가드레일' 기능의 핵심 요소인 비용 및 토큰 정보를 제공합니다.
@@ -8,7 +10,6 @@ package com.llm_ops.demo.gateway.dto;
  * @param estimatedCost API 호출에 대한 예상 비용
  */
 public record GatewayChatUsage(
-        Long totalTokens,
-        Double estimatedCost
-) {
+                Long totalTokens,
+                BigDecimal estimatedCost) {
 }
